Updated toolchain for composer repo
Changes in the upstream Satis project required a new approach to getting the
list of packages (which we use for building distribution packages).
Additionally, due to continual issues with using github URIs when building the
composer repository, we've switched to local repositories, which required a new
`satis-local.json` pointing at the local repos (`Makefile` references it); this
file also incorporates the repository renames, as well as the additional repos
under the zendframework and zfcampus organizations. Because the repositories are
on the local filesystem, `download_packages_json.php` was updated to rewrite the
URIs for the source packages to the canonical GitHub repositories.